requirements.txt
anyio==3.6.1
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.78.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
numpy==1.22.4
Pillow==9.1.1
pydantic==1.9.1
python-multipart==0.0.5
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
